fix(connections): preview and viewer URLs respect connection prefix (C-161)
Preview thumbnails on connection cards and the object viewer were
constructing S3 URLs from prefix-relative pathNames, omitting the
configured `connectionConfig.prefix`. All requests for ome.tif / zarr
returned 404 on connections with a configured prefix.

Not a SigV4 regression — signatures were fine; the keys were wrong.

Changes:
- Move `constructS3Url` from `zarrUtils.ts` to `resourceId.ts` and add
  `buildHttpsUrl(config, pathName)` — the prefix-aware wrapper that
  rejoins the configured prefix before delegating to `constructS3Url`.
- Add `httpsUrl` to `ResolvedResource` alongside `s3Key` and `s3Uri`.
- Introduce primitive selectors (`selectS3Key`, `selectHttpsUrl`,
  `selectS3Uri`). Each returns a primitive, safe to subscribe to
  reactively. Demote the composite `selectResolvedResource` (returned
  a fresh object each call — triggered re-render loops when subscribed).
- DirectoryViewGrid: prefer `_Object.Key` (absolute, from listing) and
  fall back to `selectHttpsUrl(node.id)` for nodes without `_Object`
  (e.g. recently-viewed entries reconstructed from DB records).
- objects.route: swap `constructS3Url(config, pathName)` for
  `buildHttpsUrl(config, pathName)` — same bug site, now prefix-aware.
- eslint: ignore `.claude/` (session worktree artifacts).
