+++ app/utils/connectionsStore/selectors.ts
@@ -2,7 +2,7 @@ import type { Credentials } from "@aws-sdk/client-sts";
 
 import { ConnectionsStore, useConnectionsStore } from "./useConnectionsStore";
 import type { ConnectionConfig } from "~/.generated/client";
-import { parseResourceId } from "~/utils/resourceId";
+import { buildHttpsUrl, parseResourceId } from "~/utils/resourceId";
 
 export interface ResolvedResource {
   connectionName: string;
@@ -14,6 +14,8 @@ export interface ResolvedResource {
   s3Key: string;
   /** S3 URI: `s3://bucketName/s3Key`. */
   s3Uri: string;
+  /** HTTPS URL for the object (virtual-hosted or path-style per bucket shape). */
+  httpsUrl: string;
 }
 
 export const select = {
@@ -41,10 +43,16 @@ export const selectConnectionIndex =
 /**
  * Non-reactive resolve for use in async callbacks and utility functions.
  * Reads current state snapshot — does not subscribe to changes.
+ *
  * @throws Error if the connection is not found in the store
+ *
+ * @example
+ * // Inside an async callback / non-reactive utility:
+ * const { credentials, connectionConfig, httpsUrl } = resolveResourceId(resourceId);
+ * const response = await signedFetch(httpsUrl);
  */
 export function resolveResourceId(resourceId: string): ResolvedResource {
-  const resolved = selectResolvedResource(resourceId)(useConnectionsStore.getState());
+  const resolved = resolveResource(resourceId, useConnectionsStore.getState());
   if (!resolved) {
     const { connectionName } = parseResourceId(resourceId);
     throw new Error(`No connection found for: ${connectionName}`);
@@ -53,31 +61,76 @@ export function resolveResourceId(resourceId: string): ResolvedResource {
 }
 
 /**
- * Zustand selector that resolves a resourceId (`connectionName/pathName`) into
- * the full connection record with S3 key and URI.
+ * Reactive selector: full S3 object key (prefix + pathName) for a resourceId,
+ * or `null` if the connection isn't in the store yet.
  *
- * Use in components for reactivity:
- *   `useConnectionsStore(selectResolvedResource(resourceId))`
+ * Primitive return — safe to subscribe to directly (stable `Object.is`).
  *
- * Use in async callbacks:
- *   `resolveResourceId(resourceId)` (non-reactive wrapper above)
+ * @example
+ * const s3Key = useConnectionsStore(selectS3Key(resourceId));
  */
-export const selectResolvedResource =
+export const selectS3Key =
   (resourceId: string) =>
-  (state: ConnectionsStore): ResolvedResource | null => {
-    const { connectionName, pathName } = parseResourceId(resourceId);
-    const conn = state.connections[connectionName];
-    if (!conn?.connectionConfig || !conn?.credentials) return null;
+  (state: ConnectionsStore): string | null =>
+    resolveResource(resourceId, state)?.s3Key ?? null;
 
-    const prefix = conn.connectionConfig.prefix?.replace(/\/$/, "");
-    const s3Key = prefix ? `${prefix}/${pathName}` : pathName;
+/**
+ * Reactive selector: HTTPS URL for a resourceId, or `null` if the connection
+ * isn't in the store yet. Virtual-hosted or path-style per bucket shape.
+ *
+ * Primitive return — safe to subscribe to directly (stable `Object.is`).
+ *
+ * @example
+ * const httpsUrl = useConnectionsStore(selectHttpsUrl(resourceId));
+ * // → "https://bucket.s3.eu-central-1.amazonaws.com/prefix/path/to/file.ome.tif"
+ */
+export const selectHttpsUrl =
+  (resourceId: string) =>
+  (state: ConnectionsStore): string | null =>
+    resolveResource(resourceId, state)?.httpsUrl ?? null;
 
-    return {
-      connectionName,
-      pathName,
-      connectionConfig: conn.connectionConfig,
-      credentials: conn.credentials,
-      s3Key,
-      s3Uri: `s3://${conn.connectionConfig.bucketName}/${s3Key}`,
-    };
+/**
+ * Reactive selector: `s3://bucket/key` URI for a resourceId, or `null` if the
+ * connection isn't in the store yet.
+ *
+ * Primitive return — safe to subscribe to directly (stable `Object.is`).
+ *
+ * @example
+ * const s3Uri = useConnectionsStore(selectS3Uri(resourceId));
+ */
+export const selectS3Uri =
+  (resourceId: string) =>
+  (state: ConnectionsStore): string | null =>
+    resolveResource(resourceId, state)?.s3Uri ?? null;
+
+/**
+ * Internal: resolves a resourceId against a store snapshot into the full
+ * `ResolvedResource` record. **Not safe to use as a reactive selector** — each
+ * call returns a fresh object literal, so subscribing via
+ * `useConnectionsStore(selectResolvedResource(id))` triggers a re-render on
+ * every store tick.
+ *
+ * For reactive use, subscribe to a primitive via `selectS3Key`, `selectHttpsUrl`,
+ * or `selectS3Uri`. For non-reactive use (async callbacks), use `resolveResourceId`.
+ */
+function resolveResource(
+  resourceId: string,
+  state: ConnectionsStore,
+): ResolvedResource | null {
+  const { connectionName, pathName } = parseResourceId(resourceId);
+  const conn = state.connections[connectionName];
+  if (!conn?.connectionConfig || !conn?.credentials) return null;
+
+  const prefix = conn.connectionConfig.prefix?.replace(/\/$/, "");
+  const s3Key = prefix ? `${prefix}/${pathName}` : pathName;
+
+  return {
+    connectionName,
+    pathName,
+    connectionConfig: conn.connectionConfig,
+    credentials: conn.credentials,
+    s3Key,
+    s3Uri: `s3://${conn.connectionConfig.bucketName}/${s3Key}`,
+    httpsUrl: buildHttpsUrl(conn.connectionConfig, pathName),
   };
+}
